:recycle: [Refactor] 서비스 인터페이스로 변경
